move component workflow functionality into simple_workflow.py
